refactor: Move calculate_cost function to configure.py
