+++ local_operator/model/registry.py
@@ -37,30 +37,6 @@ class ModelInfo(BaseModel):
         return value
 
 
-def calculate_cost(model_info: ModelInfo, input_tokens: int, output_tokens: int) -> float:
-    """
-    Calculates the cost of a request based on token usage and model pricing.
-
-    Args:
-        model_info (ModelInfo): The pricing information for the model.
-        input_tokens (int): The number of input tokens used in the request.
-        output_tokens (int): The number of output tokens generated by the request.
-
-    Returns:
-        float: The total cost of the request.
-
-    Raises:
-        ValueError: If there is an error during cost calculation.
-    """
-    try:
-        input_cost = (input_tokens / 1_000_000) * model_info.input_price
-        output_cost = (output_tokens / 1_000_000) * model_info.output_price
-        total_cost = input_cost + output_cost
-        return total_cost
-    except Exception as e:
-        raise ValueError(f"Error calculating cost: {e}") from e
-
-
 def get_model_info(hosting: str, model: str) -> ModelInfo:
     """
     Retrieves the model information based on the hosting provider and model name.
